refactor: chat header service status
